update to resolve CA-98: BMDB should be able to handle Biomarker Panels.
The model has been updated to give Biomarkers an IsPanel flag which is used
to signify whether or not a Biomarker is a Panel. If a Biomarker is a panel, its 
"members" can be accessed from the PanelMembers array. Both IsPanel, and the 
list of "member" Biomarkers are set by the user when the Biomarker is first 
created.